refac: MCP 서버 확장에 따른 구조 리팩토링 추가
